reverted separate renderer to better track changes
