+++ examples/example/main.rs
@@ -1115,9 +1115,9 @@ impl App {
             egui::FontDefinitions::default(),
             egui::Style::default(),
             device.clone(),
+            Arc::clone(&allocator),
             graphics_queue_index,
             graphics_queue,
-            Arc::clone(&allocator),
             swapchain_loader.clone(),
             swapchain.clone(),
             format.clone(),
@@ -1729,7 +1729,7 @@ impl App {
         }
 
         // #### egui ##########################################################################
-        self.egui_integration.resize(
+        self.egui_integration.update_swapchain(
             self.width,
             self.height,
             self.swapchain.clone(),
